+++ user/src/main/java/com/sparta/user/domain/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.sparta.user.domain.controller;
 
-import com.sparta.user.domain.dto.request.SignupRequestDto;
 import com.sparta.user.domain.dto.request.UpdateUserRequestDto;
 import com.sparta.user.domain.dto.response.UserResponseDto;
 import com.sparta.user.domain.dto.response.UsernameResponseDto;
@@ -13,7 +12,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -26,14 +24,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<UsernameResponseDto> signup(@Valid @RequestBody SignupRequestDto requestDto) {
-
-        UsernameResponseDto responseDto = userService.signup(requestDto);
-
-        return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
-    }
 
     @GetMapping("/{username}")
     public ResponseEntity<UserResponseDto> getUser(
