+++ user/src/main/java/com/sparta/user/domain/service/UserService.java
@@ -1,7 +1,6 @@
 package com.sparta.user.domain.service;
 
 import com.sparta.user.config.AuditAwareImpl;
-import com.sparta.user.domain.dto.request.SignupRequestDto;
 import com.sparta.user.domain.dto.request.UpdateUserRequestDto;
 import com.sparta.user.domain.dto.response.UserResponseDto;
 import com.sparta.user.domain.dto.response.UsernameResponseDto;
@@ -11,7 +10,6 @@ import com.sparta.user.model.entity.UserRoleEnum;
 import com.sparta.user.model.repository.UserRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,24 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
     private final AuditAwareImpl auditAware;
-
-    @Transactional
-    public UsernameResponseDto signup(@Valid SignupRequestDto requestDto) {
-        validateUsername(requestDto.getUsername());
-        validateNickname(requestDto.getNickname());
-        validateSlackId(requestDto.getSlackId());
-        validateRole(requestDto.getRole());
-
-        User user = requestDto.toEntity(passwordEncoder);
-        user.updateCreatedByAndLastModifiedBy(requestDto.getUsername());
-        userRepository.save(user);
-
-        return UsernameResponseDto.builder()
-                .username(user.getUsername())
-                .build();
-    }
 
     public UserResponseDto getUser(String requestUsername, UserRoleEnum requestRole, String username) {
         // 요청 헤더의 role이 MASTER가 아닌 경우, 자신의 정보만 확인할 수 있음
@@ -88,34 +69,6 @@ public class UserService {
         return UsernameResponseDto.builder()
                 .username(user.getUsername())
                 .build();
-    }
-
-    private void validateUsername(String username) {
-        boolean checkUsername = userRepository.existsById(username);
-        if (checkUsername) {
-            throw new IllegalArgumentException(UserExceptionMessage.DUPLICATED_USERNAME.getMessage());
-        }
-    }
-
-    private void validateNickname(String nickname) {
-        boolean checkNickname = userRepository.existsByNickname(nickname);
-        if (checkNickname) {
-            throw new IllegalArgumentException(UserExceptionMessage.DUPLICATED_NICKNAME.getMessage());
-        }
-    }
-
-    private void validateSlackId(String slackId) {
-        boolean checkSlackId = userRepository.existsBySlackId(slackId);
-        if (checkSlackId) {
-            throw new IllegalArgumentException(UserExceptionMessage.DUPLICATED_SLACKID.getMessage());
-        }
-    }
-
-    // 회원가입 시 Request Body의 role이 MASTER인지 검증하는 메서드
-    private void validateRole(UserRoleEnum role) {
-        if (role == UserRoleEnum.MASTER) {
-            throw new IllegalArgumentException(UserExceptionMessage.NOT_ALLOWED_ROLE.getMessage());
-        }
     }
 
     // 요청 헤더의 role이 MASTER인지 검증하는 메서드
